requirements.txt
pygame==2.5.2
pyserial==3.5
pynmea2==1.18.0